requirements.txt
fastapi==0.104.1
uvicorn==0.23.2

pandas==2.0.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==0.21.0
pymysql==1.1.0
sqlalchemy==2.0.23
pydantic==1.10.8